Rename all occurences of `cloak` with `dagger`
This is a significant & disruptive change, please read this message with
the care & attention so that the surprises are kept to a minimum.

"Project Cloak" was the codename for what will become Dagger v0.3.x. As
we start preparing to ship Dagger v0.3.x pre-releases, we need to use
the same "dagger" name everywhere so that:

1. Users keep the same entrypoint: the `dagger` binary

2. The tooling that installs `dagger` (e.g. HomeBrew, install.sh,
   install.ps) largely continues working as is. The only difference is
   that now the tooling will gain the functionality to install
   pre-releases of `dagger`.

There is more context here (including what happens after this gets
merged): https://github.com/dagger/dagger/issues/3176#issuecomment-1267359683
