+++ .gitignore
@@ -33,4 +33,4 @@ venv
 __pycache__
 
 # Built Binary
-/cloak
+/dagger
