--- cloak.yaml
@@ -1,5 +1,0 @@
-name: examples
-dependencies:
-  - local: examples/alpine/cloak.yaml
-  - local: examples/yarn/cloak.yaml
-  - local: examples/netlify/go/cloak.yaml
