+++ core/integration/project_test.go
@@ -16,7 +16,7 @@ import (
 func TestExtensionMount(t *testing.T) {
 	startOpts := &engine.Config{
 		Workdir:    "../../",
-		ConfigPath: "core/integration/testdata/extension/cloak.yaml",
+		ConfigPath: "core/integration/testdata/extension/dagger.yaml",
 	}
 
 	err := engine.Start(context.Background(), startOpts, func(ctx engine.Context) error {
@@ -74,8 +74,8 @@ func TestExtensionMount(t *testing.T) {
 func TestGoGenerate(t *testing.T) {
 	tmpdir := t.TempDir()
 
-	cloakYamlPath := filepath.Join(tmpdir, "cloak.yaml")
-	err := os.WriteFile(cloakYamlPath, []byte(`
+	daggerYamlPath := filepath.Join(tmpdir, "dagger.yaml")
+	err := os.WriteFile(daggerYamlPath, []byte(`
 name: testgogenerate
 scripts:
   - path: .
