+++ core/integration/testdata/extension/dagger.yaml
@@ -1,0 +1,4 @@
+name: test
+extensions:
+  - path: .
+    sdk: go
