+++ demos/v2/demo-v2.md
@@ -2,16 +2,16 @@
 
 ## Outline
 
-In this demo we show how a realistic scenario of how a software project might adopt Cloak, and why.
+In this demo we show how a realistic scenario of how a software project might adopt Dagger, and why.
 
 This demo was delivered live (with some technical issues) at [VMware Explore 2022](https://www.twitch.tv/videos/1578035654).
 
 Material to prepare this demo:
 
-* Build `cloak` from this tag: https://github.com/dagger/cloak/releases/tag/demov2
+* Build `Dagger` from this tag: https://github.com/dagger/cloak/releases/tag/demov2
 * Checkout the todoapp repository from this branch: https://github.com/dagger/todoapp/tree/cloak-demo-v2
 
-## Why use cloak/dagger?
+## Why use Dagger?
 
 TODO
