+++ docs/guides/y0yh0-writing_extensions_go.md
@@ -13,7 +13,7 @@ Say we are creating a new project called `foo`. It will have
 ## Setup the project configuration
 
 1. Enter a go module directory for your project (`go mod init <module name>` if one doesn't exist)
-1. Configure go to use the private cloak git repo (this will go away once the repo is made public)
+1. Configure go to use the `cloak` branch in the dagger git repo
 
    - Then run the following commands to setup the rest of the required dependencies
 
@@ -23,9 +23,9 @@ Say we are creating a new project called `foo`. It will have
      go mod edit -replace=github.com/docker/docker=github.com/docker/docker@v20.10.3-0.20220414164044-61404de7df1a+incompatible
      ```
 
-1. Create a new file called `cloak.yaml`
+1. Create a new file called `dagger.yaml`
 
-   - This is where you declare your project, and other project that it depends on. All extensions declared in this file will be built, loaded, and available to be called when the project is loaded into cloak.
+   - This is where you declare your project, and other project that it depends on. All extensions declared in this file will be built, loaded, and available to be called when the project is loaded into dagger.
    - Create the file in the following format:
 
    ```yaml
@@ -37,11 +37,11 @@ Say we are creating a new project called `foo`. It will have
      - git:
          remote: https://github.com/dagger/dagger.git
          ref: cloak
-         path: examples/yarn/cloak.yaml
+         path: examples/yarn/dagger.yaml
      - git:
          remote: https://github.com/dagger/dagger.git
          ref: cloak
-         path: examples/netlify/go/cloak.yaml
+         path: examples/netlify/go/dagger.yaml
    ```
 
    - The dependencies are optional and just examples, feel free to change as needed.
@@ -51,18 +51,18 @@ Say we are creating a new project called `foo`. It will have
 
 ### Generate initial script code
 
-1. From the directory containing `cloak.yaml` run `mkdir script`
-1. From the directory containing `cloak.yaml` (or any subdirectory thereof), run `cloak generate`
+1. From the directory containing `dagger.yaml` run `mkdir script`
+1. From the directory containing `dagger.yaml` (or any subdirectory thereof), run `dagger generate`
 1. You should now see:
 
    - A `script/main.go` file
 
-1. NOTE: if you re-run `cloak generate` in the future, it will refuse to overwrite your existing `main.go` file if it exists (in order to not clobber your work). For now, if you update your schema you won't get new autogenerated skeletons for it unless you temporarily rename your `main.go` file. We intend to make this less tedious with future SDK updates.
+1. NOTE: if you re-run `dagger generate` in the future, it will refuse to overwrite your existing `main.go` file if it exists (in order to not clobber your work). For now, if you update your schema you won't get new autogenerated skeletons for it unless you temporarily rename your `main.go` file. We intend to make this less tedious with future SDK updates.
 
 ### Implement the script
 
 1. Edit `script/main.go`, replacing the `panic("implement me")` with your actual implementation.
-1. When you need to call a dependency declared in `cloak.yaml`, you will currently have to use raw graphql queries. Examples of this can be found in [the alpine extension here](https://github.com/dagger/dagger/blob/cloak/examples/alpine/main.go).
+1. When you need to call a dependency declared in `dagger.yaml`, you will currently have to use raw graphql queries. Examples of this can be found in [the alpine extension here](https://github.com/dagger/dagger/blob/cloak/examples/alpine/main.go).
 1. Also feel free to import any other third-party libraries as needed the same way you would with any other go project.
 
 ### Invoke your script
@@ -72,7 +72,7 @@ Say we are creating a new project called `foo`. It will have
 
 ## Create your extension
 
-Update your `cloak.yaml` to include a new `extensions` key:
+Update your `dagger.yaml` to include a new `extensions` key:
 
 ```yaml
 name: foo
@@ -86,11 +86,11 @@ dependencies:
   - git:
       remote: https://github.com/dagger/dagger.git
       ref: cloak
-      path: examples/yarn/cloak.yaml
+      path: examples/yarn/dagger.yaml
   - git:
       remote: https://github.com/dagger/dagger.git
       ref: cloak
-      path: examples/netlify/go/cloak.yaml
+      path: examples/netlify/go/dagger.yaml
 ```
 
 ### Create schema file
@@ -116,18 +116,18 @@ dependencies:
 
 ### Generate initial extension code
 
-1. From any project directory (that is, the directory containing `cloak.yaml` or any subdirectory thereof), run `cloak generate`
+1. From any project directory (that is, the directory containing `dagger.yaml` or any subdirectory thereof), run `dagger generate`
 
 1. You should now see:
    - A `ext/main.go` file
-   - Some autogenerated boilerplate: structures for needed types in `models.go` and some runtime code that makes your extension invokable in the containerized cloak engine in `generated.go`
-1. NOTE: this has the same behavior as scripts in that `cloak generate` will refuse to overwrite any existing `main.go` file (mentioned above)
+   - Some autogenerated boilerplate: structures for needed types in `models.go` and some runtime code that makes your extension invokable in the containerized dagger engine in `generated.go`
+1. NOTE: this has the same behavior as scripts in that `dagger generate` will refuse to overwrite any existing `main.go` file (mentioned above)
 
 ### Implement the extension
 
 1. Edit `ext/main.go`, replacing occurences of `panic("implement me")` with the implementation of your extension's actions.
-1. When you need to call a dependency declared in `cloak.yaml`, you will currently have to use raw graphql queries. Examples of this can be found in [the alpine extension here](https://github.com/dagger/dagger/blob/cloak/examples/alpine/main.go).
-1. Also feel free to import any other third-party dependencies as needed the same way you would with any other go project. They should all be installed and available when executing in the cloak engine.
+1. When you need to call a dependency declared in `dagger.yaml`, you will currently have to use raw graphql queries. Examples of this can be found in [the alpine extension here](https://github.com/dagger/dagger/blob/cloak/examples/alpine/main.go).
+1. Also feel free to import any other third-party dependencies as needed the same way you would with any other go project. They should all be installed and available when executing in the dagger engine.
 1. Some examples:
    - [alpine](https://github.com/dagger/dagger/blob/cloak/examples/alpine/main.go)
    - [netlify](https://github.com/dagger/dagger/blob/cloak/examples/netlify/go/main.go)
@@ -135,14 +135,14 @@ dependencies:
 ### Invoke your extension
 
 1. One simple way to verify your extension builds and can be invoked is via the graphql playground.
-   - Just run `cloak dev` from any directory in your project and navigate to `localhost:8080` in your browser (may need [an SSH tunnel](https://www.ssh.com/academy/ssh/tunneling-example) if on a remote host)
+   - Just run `dagger dev` from any directory in your project and navigate to `localhost:8080` in your browser (may need [an SSH tunnel](https://www.ssh.com/academy/ssh/tunneling-example) if on a remote host)
      - you can use the `--port` flag to override the port if needed
    - Click the "Docs" tab on the right to see the schemas available, including your extension and any dependencies.
    - You can submit queries by writing them on the left-side pane and clicking the play button in the middle
-1. You can also use the cloak CLI, e.g.
+1. You can also use the dagger CLI, e.g.
 
    ```console
-   cloak do <<'EOF'
+   dagger do <<'EOF'
    {
      foo {
        bar(in: "in")
