--- examples/alpine/cloak.yaml
@@ -1,4 +1,0 @@
-name: alpine
-extensions:
-  - path: .
-    sdk: go
