--- examples/gitdeps/cloak.yaml
@@ -1,14 +1,0 @@
-name: gitdeps-example
-dependencies:
-  - git:
-      remote: https://github.com/dagger/dagger.git
-      ref: cloak
-      path: examples/alpine/cloak.yaml
-  - git:
-      remote: https://github.com/dagger/dagger.git
-      ref: cloak
-      path: examples/yarn/cloak.yaml
-  - git:
-      remote: https://github.com/dagger/dagger.git
-      ref: cloak
-      path: examples/netlify/go/cloak.yaml
