+++ examples/netlify/ts/dagger.yaml
@@ -1,0 +1,4 @@
+name: netlify
+extensions:
+  - path: .
+    sdk: ts
