--- examples/python/hello/cloak.yaml
@@ -1,4 +1,0 @@
-name: "python"
-extensions:
-  - path: .
-    sdk: python
