+++ examples/python/hello/dagger.yaml
@@ -1,0 +1,4 @@
+name: "python"
+extensions:
+  - path: .
+    sdk: python
