--- examples/yarn/cloak.yaml
@@ -1,6 +1,0 @@
-name: yarn
-extensions:
-  - path: .
-    sdk: ts
-dependencies:
-  - local: ../alpine/cloak.yaml
