+++ examples/yarn/dagger.yaml
@@ -1,0 +1,6 @@
+name: yarn
+extensions:
+  - path: .
+    sdk: ts
+dependencies:
+  - local: ../alpine/dagger.yaml
