+++ project/sdk.go
@@ -9,7 +9,7 @@ import (
 
 // TODO:(sipsma) SDKs should be pluggable extensions, not hardcoded LLB here. The implementation here is a temporary bridge from the previous hardcoded Dockerfiles to the sdk-as-extension model.
 
-// return the FS with the executable extension code, ready to be invoked by cloak
+// return the FS with the executable extension code, ready to be invoked by dagger
 func (s RemoteSchema) Runtime(ctx context.Context, ext *Extension) (*filesystem.Filesystem, error) {
 	var runtimeFS *filesystem.Filesystem
 	var err error
